# Optimize GetControllerName for CNP