--- a/pkg/k8s/apis/cilium.io/v2/cnp_types.go
+++ b/pkg/k8s/apis/cilium.io/v2/cnp_types.go
@@ -6,6 +6,7 @@ package v2
 import (
 	"fmt"
 	"reflect"
+	"strings"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -238,7 +239,14 @@ func (r *CiliumNetworkPolicy) Parse() (api.Rules, error) {
 // GetControllerName returns the unique name for the controller manager.
 func (r *CiliumNetworkPolicy) GetControllerName() string {
 	name := k8sUtils.GetObjNamespaceName(&r.ObjectMeta)
-	return fmt.Sprintf("%s (v2 %s)", k8sConst.CtrlPrefixPolicyStatus, name)
+	const staticLen = 6
+	var str strings.Builder
+	str.Grow(staticLen + len(name) + len(k8sConst.CtrlPrefixPolicyStatus))
+	str.WriteString(k8sConst.CtrlPrefixPolicyStatus)
+	str.WriteString(" (v2 ")
+	str.WriteString(name)
+	str.WriteString(")")
+	return str.String()
 }
 
 // GetIdentityLabels returns all rule labels in the CiliumNetworkPolicy.