--- a/pkg/k8s/apis/cilium.io/v2/cnp_types_test.go
+++ b/pkg/k8s/apis/cilium.io/v2/cnp_types_test.go
@@ -1,0 +1,33 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package v2
+
+import (
+	"testing"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func BenchmarkCNPGetControllerName(b *testing.B) {
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for _, s := range []CiliumNetworkPolicy{
+			{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "default",
+				},
+			},
+			{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "longname",
+					Namespace: "",
+				},
+			},
+		} {
+			_ = s.GetControllerName()
+		}
+	}
+}