Use strings.Builder instead of fmt.Sprintf().

Results:
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkCNPGetControllerName-10     338           206           -39.05%

benchmark                            old allocs     new allocs     delta
BenchmarkCNPGetControllerName-10     6              4              -33.33%

benchmark                            old bytes     new bytes     delta
BenchmarkCNPGetControllerName-10     464           432           -6.90%
```

Signed-off-by: Marcel Zięba <marseel@gmail.com>
Related: https://github.com/cilium/cilium/issues/19571